move context construction duties from wasi-cap-std-sync to wasmtime-wasi-sockets-sync
